AgroAISearch/FrontEnd/src/requirements.txt
azure-cognitiveservices-speech
streamlit-mic-recorder
sentence_transformers
InstructorEmbedding
langchain==0.0.341
ctransformers
transformers
streamlit
pypdfium2
chromadb
librosa
pyyaml
protobuf==3.20.3